@@ -6,28 +6,25 @@ import (
 	"github.com/GongfuTea/gft-go/types"
 )
 
-type MgoTreeRepo struct {
-	*MgoRepo
+type MgoTreeRepo[T types.ITreeEntity] struct {
+	*MgoRepo[T]
 }
 
-func NewMgoTreeRepo(name string, factory func() types.IEntity) *MgoTreeRepo {
-	return &MgoTreeRepo{
-		&MgoRepo{
+func NewMgoTreeRepo[T types.ITreeEntity](name string, factory func() T) *MgoTreeRepo[T] {
+	return &MgoTreeRepo[T]{
+		&MgoRepo[T]{
 			Name: name, factory: factory,
 		},
 	}
 }
 
-func (repo MgoTreeRepo) Save(m types.IEntity) (types.IEntity, error) {
-	fmt.Printf("save MgoTreeRepo, %#v\n", m)
+func (repo MgoTreeRepo[T]) Save(model T) (T, error) {
+	fmt.Printf("save MgoTreeRepo, %#v\n", model)
 
-	model, _ := m.(types.ITreeEntity)
 	if model.HasPid() {
 		parent, err := repo.MgoRepo.Get(model.PID())
 		if err == nil {
-			if tree, ok := parent.(types.ITreeEntity); ok {
-				model.SetMpath(tree.GetMpath() + model.GetCode() + ".")
-			}
+			model.SetMpath(parent.GetMpath() + model.GetCode() + ".")
 		}
 	} else {
 		fmt.Printf("save category, %#v\n", model)
