@@ -1,24 +1,23 @@
-package base
+package jc
 
 import (
 	"github.com/GongfuTea/gft-go/core/db"
 )
 
-type GftDictItem struct {
+type GftGsZydm struct {
 	*db.DbEntity `bson:",inline"`
-	CategoryId   string  `bson:"categoryId" json:"categoryId"`
 	Code         string  `bson:"code" json:"code"`
 	Name         string  `bson:"name" json:"name"`
-	Nickname     string  `bson:"nickname" json:"nickname"`
 	SortOrder    float32 `bson:"sortOrder" json:"sortOrder"`
 	Level        int     `bson:"level" json:"level"`
 	Note         string  `bson:"note" json:"note"`
+	Xwlbm        string  `bson:"xwlbm" json:"xwlbm"` // 学位类别
+	Xkmlm        string  `bson:"xkmlm" json:"xkmlm"` // 学科门类
 
-	Locale map[string]string `bson:"locale,omitempty" json:"locale,omitempty"`
 }
 
-func NewGftDictItem() *GftDictItem {
-	return &GftDictItem{
+func NewGftGsZydm() *GftGsZydm {
+	return &GftGsZydm{
 		DbEntity: db.NewDbEntity(),
 	}
 }
